Add files via upload
I created this algorithm for creating weekly schedule for businesses which run with the part time employee. It allows the user to manually enter number of employee, available dates of each employee for a week, number of days each employee would like to work, and the priority( you can also set all the employee to the same priority level if you do not need to give any privilege). The whole algorithm is created using purely C. So, there is no user interface and you need to manually type in all the information into the array provided.