Add files via upload
A small sorting and searching algorithm using C++. At first the program ask the user to enter an array of integers of desired size and then it sort the array using the bubble sort algorithm and print it out, and finally the program ask the user to enter the number he/she would like to search for. The program use a simple binary search algorithm to search for the number the user has entered and print out the number's index in the sorted array.